// apps/api/src/modules/inventory/entities/inventory-item.entity.d.ts
import { Brand } from './brand.entity';
import { Model } from './model.entity';
import { Branch, User } from '../../auth/entities/user.entity';
import { ItemPhoto } from './item-photo.entity';
export declare class InventoryItem {
    id: string;
    imei: string;
    imei2: string;
    brand: Brand;
    brandId: string;
    model: Model;
    modelId: string;
    colour: string;
    storage: string;
    ram: string;
    boxType: string;
    pkuCode: string;
    batteryHealth: number;
    countryOfOrigin: string;
    hsnCode: string;
    condition: string;
    itemName: string;
    firstInvoiceDate: Date;
    purchasePrice: number;
    wholesalePrice: number;
    boxPrice: number;
    taxRate: number;
    taxAmount: number;
    totalCost: number;
    sellingPrice: number;
    onlinePrice: number;
    status: string;
    isOnline: boolean;
    birthdayOffer: boolean;
    branch: Branch;
    branchId: string;
    purchaseId: string | null;
    notes: string;
    accessories: object;
    images: string[];
    warrantyStatus: string;
    warrantyExpiry: Date | null;
    createdBy: User;
    createdById: string;
    photos: ItemPhoto[];
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=inventory-item.entity.d.ts.map